Database up and running
The database is up  and running — I’ve written a schema and some
boilerplate code to create the database, and I’ve also installed a
Swift SQLite wrapper, since there is no Swift interface to SQLite. I’ve
written a couple simple queries that work as a proof of concept.