Added Music and Sound Effects (Not fully complete)
I added the background music for Space, Beach, Underwater, and Desert. I
also added a Twitter logo and Facebook logo in case we want to use that
functionality in the future since Eric said it was pretty easy to do. I
also added in the original boards folder sound effects for the puck
hitting walls, the paddles hitting the puck as well as when a goal is
scored. (it involved some youtube ripping and editing of mp3 files...
good times XD) I also added menu music. It was from a pretty popular
song but it is instrumental and I can imagine how awesome it will fit
with the menus :D If you guys want to change it later we can just let me
know.